updated config file, cmorizer, utilities, check_obs recipe
